Completed JSON grabber and ArgumentParser
Completed get_json.py: Module that will take user inputted json files/URL and returns a json object that is readable by the rest of the program.

Completed the standard ArgumentParser in main that defines the usage of the program (see main.py).

Added free.py as a freestanding random test file.